Stem Center Analytics Public Code Debut
While the project is still a few months away from a polished product, I
am proud to anounce that we have our first public debut of the Foothill
College's Stem Center Analytics source code!

Despite this timing of this debut, this project has been under works for
over a year and a half now in a private repo. This was primarily due to
the state of the data, as it was not anonymized for public consumption).