Fixed bug in documentation, added error supression decorator
There was a bug in the docs caused by a refactoring issue where all was
changed to ALL_SUBJECTS, which managed to sneek in to a recent commit by
mistake. All the instances of this were reverted back to the original
(ALL_SUBJECTS -> all) for all modules in the branch.

Also, a custom made decorator that adds the ability to suprress
exceptions of decorated functions was added to
stem_center_analytics.os_lib. In a future commit this will allow certain
os related functions (like remove file) to work silently and not report
the error (useful in a few cases in this project).