Implemented data retrieval script for imap server
Wrote a script (and verified that it works) that downloads the
attachments of the first unread email, marking it as read, and starring
it if an error occurs.

For obvious reasons, this is done through a secured email account
elsewhere. The email server is setup to only recieve and download
specific messages from specific sources (the data source's email
address).