Overrode unclean tutor requests with improved columns
Uploaded tutor requests with the added distinction of time of request
and time of service, in which to-the-second wait-times can be
calculated.

This is part of my recent initiative to minimize the dependency on the
SARS stem center data sources. This means that as much that can be
inferred and calculated from the data will be done in the backend
instead, due to the unreliability of the contracted service. This is all
part of the steps toward a fully functioning, maximally automated
system.

DISCLAIMER: The database nor the rest of the project has yet to
accommodate these changes.