Simplified Deployment
Simplified deployment by completely removing the need for private/public
repos (at least for now) - as the sensitive information is now set as
config variables neatly seperated from file content.

This renders the shellscripts unnecessary as their is no need for
overriding git ignores or force adding files depending on the repo. This
means all the content on the heroku server is essentially the same as
the content on the public repo -- from now on.